finally displays again... kinda
